- PLUGIN_safe_assets: logout actions can be (un)set, for speed optimization
git-svn-id: file:///Users/jan/Dropbox/svn/FlexyAdmin/trunk@3671 0c02b2c5-a8bd-4e35-85b2-cf8cacb96951

Former-commit-id: 327566db1b120c6a8ba474fb04773d0d65bfc6fc
Former-commit-id: 4ae559d3a45efe430acadf74f92c4332ab7036bf
Former-commit-id: b67025ce4c7bc77bc599400f0cb42a773009e4bd [formerly c509ca9acd946fdf5b66b5a0deecf3705575e0a7]
Former-commit-id: 8ed747365080f3a578664e0e3242c0eef3c68427

+++ sys/flexyadmin/libraries/plugins/Plugin_safe_assets.php
@@ -92,14 +92,18 @@ class Plugin_safe_assets extends Plugin {
    */
   function _admin_logout() {
 		$logout=true;
-    // Cleanup captcha's
-    $files = glob('site/assets/_thumbcache/captcha*');
-    if (is_array($files)) {
-      foreach( $files as $file ) {
-        unlink($file);
+    $logout_actions = $this->config('logout_actions');
+    if ( el('cleanup_captha',$logout_actions,false) ) {
+      $files = glob('site/assets/_thumbcache/captcha*');
+      if (is_array($files)) {
+        foreach( $files as $file ) {
+          unlink($file);
+        }
       }
     }
-		$this->_safe_and_clean_all();
+    if ( el('clean_all',$logout_actions,false) ) {
+      $this->_safe_and_clean_all();
+    }
     if (!empty($this->removed)) return $this->_show();
     return FALSE;
 	}
